requirements.txt
google-api-python-client
yt-dlp
requests
python-dotenv